Web: Files: moved advancedOptions to AccessComboBox
